# output/final_deliverables/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "N100 Financial Intelligence API",
    "description": "REST API exposing financial analytics, screening, valuation, peer comparison, portfolio statistics, and company reports for the Nifty 100 universe.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/screen": {
      "get": {
        "tags": [
          "Screener"
        ],
        "summary": "Screen Companies",
        "description": "Executes an investment screener preset and returns matching companies.\nIf no preset is provided, returns the list of available presets.",
        "operationId": "screen_companies_api_v1_screen_get",
        "parameters": [
          {
            "name": "preset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Preset strategy name (e.g. 'Value Pick')",
              "title": "Preset"
            },
            "description": "Preset strategy name (e.g. 'Value Pick')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/screener": {
      "get": {
        "tags": [
          "Screener"
        ],
        "summary": "Get Screener",
        "description": "Filters and screens companies based on latest-year KPI metrics.",
        "operationId": "get_screener_api_v1_screener_get",
        "parameters": [
          {
            "name": "min_roe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum ROE %",
              "title": "Min Roe"
            },
            "description": "Minimum ROE %"
          },
          {
            "name": "max_de",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum Debt to Equity ratio",
              "title": "Max De"
            },
            "description": "Maximum Debt to Equity ratio"
          },
          {
            "name": "min_fcf",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum Free Cash Flow (Cr)",
              "title": "Min Fcf"
            },
            "description": "Minimum Free Cash Flow (Cr)"
          },
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Standardized sector name",
              "title": "Sector"
            },
            "description": "Standardized sector name"
          },
          {
            "name": "min_rev_cagr_5yr",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum 5Y Revenue CAGR %",
              "title": "Min Rev Cagr 5Yr"
            },
            "description": "Minimum 5Y Revenue CAGR %"
          },
          {
            "name": "min_pat_cagr_5yr",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum 5Y PAT CAGR %",
              "title": "Min Pat Cagr 5Yr"
            },
            "description": "Minimum 5Y PAT CAGR %"
          },
          {
            "name": "max_pe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum PE ratio",
              "title": "Max Pe"
            },
            "description": "Maximum PE ratio"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sector": {
      "get": {
        "tags": [
          "Sectors"
        ],
        "summary": "Get Sector Info",
        "description": "Legacy route to return statistics and list of companies for a given sector name.\nIf no sector name is provided, returns summary stats for all sectors.",
        "operationId": "get_sector_info_api_v1_sector_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sector name (e.g. 'IT')",
              "title": "Name"
            },
            "description": "Sector name (e.g. 'IT')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/portfolio/analyze": {
      "post": {
        "tags": [
          "Portfolio"
        ],
        "summary": "Analyze Portfolio",
        "description": "Computes portfolio metrics: Expected Return, Volatility, Sharpe Ratio, Beta, and Diversification.",
        "operationId": "analyze_portfolio_api_v1_portfolio_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sector/{sector_name}": {
      "get": {
        "summary": "Get Sector Data",
        "description": "Legacy route to fetch sector details (delegates to sectors router).",
        "operationId": "get_sector_data_api_sector__sector_name__get",
        "parameters": [
          {
            "name": "sector_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sector Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/screener/{preset_name}": {
      "get": {
        "summary": "Run Screener Preset",
        "description": "Legacy route to run screener presets (delegates to screener router).",
        "operationId": "run_screener_preset_api_screener__preset_name__get",
        "parameters": [
          {
            "name": "preset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PortfolioAnalysisRequest": {
        "properties": {
          "allocations": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Allocations",
            "description": "Dictionary mapping company ticker (keys) to their weight/allocation (values)",
            "example": {
              "ABB": 0.2,
              "INFY": 0.4,
              "TCS": 0.4
            }
          },
          "risk_free_rate": {
            "type": "number",
            "title": "Risk Free Rate",
            "description": "Risk-free rate in percentage points",
            "default": 7.0,
            "example": 7.0
          }
        },
        "type": "object",
        "required": [
          "allocations"
        ],
        "title": "PortfolioAnalysisRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}